libc/baselibc: Fix size overflow compiler error
On some compilers (for example 64-bit gcc 9.4.0) passing
SIZE_MAX/2 to vsnprintf as buffer size was causing an compilation error.
INT_MAX/2 should also be enough for buffer size and now it compiles
without problems.
This also removes force inclusion of "main" symbol.

+++ libc/baselibc/src/sprintf.c
@@ -4,6 +4,7 @@
 
 #include <stdio.h>
 #include <stdint.h>
+#include <limits.h>
 
 int sprintf(char *buffer, const char *format, ...)
 {
@@ -11,7 +12,7 @@ int sprintf(char *buffer, const char *format, ...)
 	int rv;
 
 	va_start(ap, format);
-	rv = vsnprintf(buffer, SIZE_MAX/2, format, ap);
+	rv = vsnprintf(buffer, INT_MAX/2, format, ap);
 	va_end(ap);
 
 	return rv;
